Move hooks to trade module
