+++ src/cow-react/common/hooks/useAreThereTokensWithSameSymbol.ts
