+++ src/cow-react/modules/limitOrders/containers/LimitOrdersWidget/index.tsx
@@ -8,10 +8,8 @@ import React, { useCallback, useState } from 'react'
 import { BalanceAndSubsidy } from 'hooks/useCowBalanceAndSubsidy'
 import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/typings'
 import { useLimitOrdersTradeState } from '../../hooks/useLimitOrdersTradeState'
-import { useSetupTradeState } from '../../hooks/useSetupTradeState'
 import { limitOrdersAtom, updateLimitOrdersAtom } from '../../state/limitOrdersAtom'
 import { useOnCurrencySelection } from '../../hooks/useOnCurrencySelection'
-import { useTradeNavigate } from '../../hooks/useTradeNavigate'
 import { useAtomValue, useUpdateAtom } from 'jotai/utils'
 import { SettingsWidget } from '../SettingsWidget'
 import { limitOrdersSettingsAtom } from '../../state/limitOrdersSettingsAtom'
@@ -25,6 +23,8 @@ import { useIsSellOrder } from '../../hooks/useIsSellOrder'
 import { TradeButtons } from '@cow/modules/limitOrders/containers/TradeButtons'
 import { TradeApproveWidget } from '@cow/common/containers/TradeApprove/TradeApproveWidget'
 import { Routes } from '@cow/constants/routes'
+import { useSetupTradeState } from '@cow/modules/trade/hooks/useSetupTradeState'
+import { useTradeNavigate } from '@cow/modules/trade/hooks/useTradeNavigate'
 
 export function LimitOrdersWidget() {
   const state = useAtomValue(limitOrdersAtom)
