+++ src/cow-react/modules/mainMenu/index.ts
@@ -1,1 +1,3 @@
 export { MainMenu } from './containers/MainMenu'
+export * from './state/mainMenuUrlOverridesAtom'
+export * from './constants/mainMenu'
