+++ src/cow-react/modules/trade/types/TradeState.ts
