+++ src/cow-react/modules/trade/updaters/TradeMenuUpdater/index.tsx
@@ -3,7 +3,7 @@ import { useAtomValue } from 'jotai/utils'
 import { Routes } from '@cow/constants/routes'
 import { MainMenuItemId } from '@cow/modules/mainMenu/constants/mainMenu'
 import { useSwapState } from 'state/swap/hooks'
-import { useParameterizeTradeInMenu } from '@cow/modules/limitOrders/hooks/useParameterizeTradeInMenu'
+import { useParameterizeTradeInMenu } from '../../hooks/useParameterizeTradeInMenu'
 
 export function TradeMenuUpdater() {
   const limitOrdersState = useAtomValue(limitOrdersAtom)
